api devlopment

+++ controllers/api/v1/posts_api.js
@@ -23,30 +23,19 @@ module.exports.destroy = async function(req,res){
     try {
             let post = await Post.findById(req.params.id);
 
-            // if(post.user == req.user.id){
+            if(post.user == req.user.id){
                 post.remove();
             
                 await Comment.deleteMany({post:req.params.id});
 
-                // if(req.xhr){
-                //     return res.status(200).json({
-                //         data:{
-                //             post_id:req.params.id
-                //         },
-                //         message:"Post Deleted"
-                //     })
-                // } 
-
-
-                // req.flash('success','Post and Comments Deleted!');
-
                 return res.json(200,{
                     message:"Post and associated comments deleted successfully!"
                 })
-        //     }else{
-        //         req.flash('error','You cannot delete this post');
-        //         return res.redirect('back');
-        // }
+            }else{
+                return res.json(401,{
+                    message:'You cannot delete this post'
+                });
+        }
     } catch (err) {
         console.log("*****",err)
         // req.flash('error',err);
